Use instances without first introducing many new type synonyms.
Define `Preorder.Antisymmetrization` to be the antisymmetrization with
respect to ≤ for a preorder. (This could later be generalised to just
an `LE` instance if necessary, but the point of antisymmetrization is
primarily for preorders.)

Construct the ordered group structures on this type, letting whichever
instances are valid get inferred, rather than defining a new
`Antisymmetrization` type for each kind of order and constructing the
instance on that.

Similarly, construct ordered structures on left/right coset spaces
directly, instead of first defining new `Antisymmetrization` type
synonyms for them.

Additionally, extract lemas related to `Equiv` to a file 'Util/Equiv.lean'.